requirements.txt
chess @ file:///home/conda/feedstock_root/build_artifacts/chess_1728077868589/work
Cython==3.0.11
filelock @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/filelock_1701804468261/work
fsspec @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_436le7786l/croot/fsspec_1724855598709/work
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1725379848064/work
imagenet-stubs @ git+https://github.com/nottombrown/imagenet_stubs@0b501276f54cbf45b8e1b67dccbdcbe51de88365
Jinja2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b15nuwux5r/croot/jinja2_1730902833938/work
keyboard==0.13.5
MarkupSafe @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/markupsafe_1707339878470/work
MouseInfo==0.1.3
mpmath @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/mpmath_1699242500508/work
networkx @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_70xddlxyrl/croot/networkx_1720002492248/work
numpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a51i_mbs7m/croot/numpy_and_numpy_base_1708638620867/work/dist/numpy-1.26.4-cp312-cp312-macosx_11_0_arm64.whl#sha256=37afb6b734a197702d848df93bd67c10b52f6467d56e518950d84b6b1c949d27
opencv-python==4.10.0
opencv-python-headless==4.10.0
pillow==11.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc==10.3.2
pyobjc-core==10.3.2
pyobjc-framework-Accessibility==10.3.2
pyobjc-framework-Accounts==10.3.2
pyobjc-framework-AddressBook==10.3.2
pyobjc-framework-AdServices==10.3.2
pyobjc-framework-AdSupport==10.3.2
pyobjc-framework-AppleScriptKit==10.3.2
pyobjc-framework-AppleScriptObjC==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-AppTrackingTransparency==10.3.2
pyobjc-framework-AudioVideoBridging==10.3.2
pyobjc-framework-AuthenticationServices==10.3.2
pyobjc-framework-AutomaticAssessmentConfiguration==10.3.2
pyobjc-framework-Automator==10.3.2
pyobjc-framework-AVFoundation==10.3.2
pyobjc-framework-AVKit==10.3.2
pyobjc-framework-AVRouting==10.3.2
pyobjc-framework-BackgroundAssets==10.3.2
pyobjc-framework-BrowserEngineKit==10.3.2
pyobjc-framework-BusinessChat==10.3.2
pyobjc-framework-CalendarStore==10.3.2
pyobjc-framework-CallKit==10.3.2
pyobjc-framework-CFNetwork==10.3.2
pyobjc-framework-Cinematic==10.3.2
pyobjc-framework-ClassKit==10.3.2
pyobjc-framework-CloudKit==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Collaboration==10.3.2
pyobjc-framework-ColorSync==10.3.2
pyobjc-framework-Contacts==10.3.2
pyobjc-framework-ContactsUI==10.3.2
pyobjc-framework-CoreAudio==10.3.2
pyobjc-framework-CoreAudioKit==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-CoreData==10.3.2
pyobjc-framework-CoreHaptics==10.3.2
pyobjc-framework-CoreLocation==10.3.2
pyobjc-framework-CoreMedia==10.3.2
pyobjc-framework-CoreMediaIO==10.3.2
pyobjc-framework-CoreMIDI==10.3.2
pyobjc-framework-CoreML==10.3.2
pyobjc-framework-CoreMotion==10.3.2
pyobjc-framework-CoreServices==10.3.2
pyobjc-framework-CoreSpotlight==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-CoreWLAN==10.3.2
pyobjc-framework-CryptoTokenKit==10.3.2
pyobjc-framework-DataDetection==10.3.2
pyobjc-framework-DeviceCheck==10.3.2
pyobjc-framework-DictionaryServices==10.3.2
pyobjc-framework-DiscRecording==10.3.2
pyobjc-framework-DiscRecordingUI==10.3.2
pyobjc-framework-DiskArbitration==10.3.2
pyobjc-framework-DVDPlayback==10.3.2
pyobjc-framework-EventKit==10.3.2
pyobjc-framework-ExceptionHandling==10.3.2
pyobjc-framework-ExecutionPolicy==10.3.2
pyobjc-framework-ExtensionKit==10.3.2
pyobjc-framework-ExternalAccessory==10.3.2
pyobjc-framework-FileProvider==10.3.2
pyobjc-framework-FileProviderUI==10.3.2
pyobjc-framework-FinderSync==10.3.2
pyobjc-framework-FSEvents==10.3.2
pyobjc-framework-GameCenter==10.3.2
pyobjc-framework-GameController==10.3.2
pyobjc-framework-GameKit==10.3.2
pyobjc-framework-GameplayKit==10.3.2
pyobjc-framework-HealthKit==10.3.2
pyobjc-framework-ImageCaptureCore==10.3.2
pyobjc-framework-InputMethodKit==10.3.2
pyobjc-framework-InstallerPlugins==10.3.2
pyobjc-framework-InstantMessage==10.3.2
pyobjc-framework-Intents==10.3.2
pyobjc-framework-IntentsUI==10.3.2
pyobjc-framework-IOBluetooth==10.3.2
pyobjc-framework-IOBluetoothUI==10.3.2
pyobjc-framework-IOSurface==10.3.2
pyobjc-framework-iTunesLibrary==10.3.2
pyobjc-framework-KernelManagement==10.3.2
pyobjc-framework-LatentSemanticMapping==10.3.2
pyobjc-framework-LaunchServices==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyobjc-framework-libxpc==10.3.2
pyobjc-framework-LinkPresentation==10.3.2
pyobjc-framework-LocalAuthentication==10.3.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.3.2
pyobjc-framework-MailKit==10.3.2
pyobjc-framework-MapKit==10.3.2
pyobjc-framework-MediaAccessibility==10.3.2
pyobjc-framework-MediaLibrary==10.3.2
pyobjc-framework-MediaPlayer==10.3.2
pyobjc-framework-MediaToolbox==10.3.2
pyobjc-framework-Metal==10.3.2
pyobjc-framework-MetalFX==10.3.2
pyobjc-framework-MetalKit==10.3.2
pyobjc-framework-MetalPerformanceShaders==10.3.2
pyobjc-framework-MetalPerformanceShadersGraph==10.3.2
pyobjc-framework-MetricKit==10.3.2
pyobjc-framework-MLCompute==10.3.2
pyobjc-framework-ModelIO==10.3.2
pyobjc-framework-MultipeerConnectivity==10.3.2
pyobjc-framework-NaturalLanguage==10.3.2
pyobjc-framework-NetFS==10.3.2
pyobjc-framework-Network==10.3.2
pyobjc-framework-NetworkExtension==10.3.2
pyobjc-framework-NotificationCenter==10.3.2
pyobjc-framework-OpenDirectory==10.3.2
pyobjc-framework-OSAKit==10.3.2
pyobjc-framework-OSLog==10.3.2
pyobjc-framework-PassKit==10.3.2
pyobjc-framework-PencilKit==10.3.2
pyobjc-framework-PHASE==10.3.2
pyobjc-framework-Photos==10.3.2
pyobjc-framework-PhotosUI==10.3.2
pyobjc-framework-PreferencePanes==10.3.2
pyobjc-framework-PushKit==10.3.2
pyobjc-framework-Quartz==10.3.2
pyobjc-framework-QuickLookThumbnailing==10.3.2
pyobjc-framework-ReplayKit==10.3.2
pyobjc-framework-SafariServices==10.3.2
pyobjc-framework-SafetyKit==10.3.2
pyobjc-framework-SceneKit==10.3.2
pyobjc-framework-ScreenCaptureKit==10.3.2
pyobjc-framework-ScreenSaver==10.3.2
pyobjc-framework-ScreenTime==10.3.2
pyobjc-framework-ScriptingBridge==10.3.2
pyobjc-framework-SearchKit==10.3.2
pyobjc-framework-Security==10.3.2
pyobjc-framework-SecurityFoundation==10.3.2
pyobjc-framework-SecurityInterface==10.3.2
pyobjc-framework-SensitiveContentAnalysis==10.3.2
pyobjc-framework-ServiceManagement==10.3.2
pyobjc-framework-SharedWithYou==10.3.2
pyobjc-framework-SharedWithYouCore==10.3.2
pyobjc-framework-ShazamKit==10.3.2
pyobjc-framework-Social==10.3.2
pyobjc-framework-SoundAnalysis==10.3.2
pyobjc-framework-Speech==10.3.2
pyobjc-framework-SpriteKit==10.3.2
pyobjc-framework-StoreKit==10.3.2
pyobjc-framework-Symbols==10.3.2
pyobjc-framework-SyncServices==10.3.2
pyobjc-framework-SystemConfiguration==10.3.2
pyobjc-framework-SystemExtensions==10.3.2
pyobjc-framework-ThreadNetwork==10.3.2
pyobjc-framework-UniformTypeIdentifiers==10.3.2
pyobjc-framework-UserNotifications==10.3.2
pyobjc-framework-UserNotificationsUI==10.3.2
pyobjc-framework-VideoSubscriberAccount==10.3.2
pyobjc-framework-VideoToolbox==10.3.2
pyobjc-framework-Virtualization==10.3.2
pyobjc-framework-Vision==10.3.2
pyobjc-framework-WebKit==10.3.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
rubicon-objc==0.4.9
screeninfo==0.8.1
setuptools==75.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1727529700862/work
torch @ file:///Users/runner/miniforge3/conda-bld/libtorch_1730862409001/work
torchinfo @ file:///home/conda/feedstock_root/build_artifacts/torchinfo_1684097133203/work
torchvision @ file:///Users/runner/miniforge3/conda-bld/torchvision_1730803229494/work
tqdm==4.67.1
typing_extensions @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_93dg13ilv4/croot/typing_extensions_1715268840722/work
wheel==0.44.0